augips/scrapers/autozone.py
"""
AutoZone store location scraper
"""

import time
from typing import List, Dict, Any, Optional
from playwright.sync_api import sync_playwright
from bs4 import BeautifulSoup
import requests

from .base import Scraper


class AutoZoneScraper(Scraper):
    """Scraper for AutoZone store locations"""
    
    def __init__(self):
        super().__init__("AutoZone")
        self.base_url = "https://www.autozone.com/locations/"
        self.store_locator_url = "https://www.autozone.com/store-locator"
    
    def scrape(self) -> List[Dict[str, Any]]:
        """
        Scrape AutoZone store locations
        
        This is an example implementation that demonstrates different scraping approaches:
        1. Static HTML parsing with BeautifulSoup
        2. Dynamic content with Playwright
        3. Geocoding fallback
        
        Returns:
            List of dictionaries containing store location data
        """
        # Example implementation - in a real scenario, choose the appropriate method
        # based on the website structure
        
        # Method 1: Static HTML parsing (placeholder)
        # locations = self._scrape_static_html()
        
        # Method 2: Dynamic content with Playwright (placeholder)
        locations = self._scrape_with_playwright()
        
        # Add company name to all locations
        for location in locations:
            location["company_name"] = self.company_name
            
            # Geocode addresses that don't have coordinates
            if "latitude" not in location or "longitude" not in location:
                address = f"{location.get('address', '')}, {location.get('city', '')}, {location.get('state', '')} {location.get('zip_code', '')}"
                lat, lng = self.geocode_address(address)
                location["latitude"] = lat
                location["longitude"] = lng
        
        return locations
    
    def _scrape_static_html(self) -> List[Dict[str, Any]]:
        """Scrape store locations from static HTML"""
        locations = []
        
        # Example implementation using requests and BeautifulSoup
        response = requests.get(self.base_url)
        if response.status_code == 200:
            soup = BeautifulSoup(response.text, "html.parser")
            
            # Find store location elements (placeholder logic)
            # In a real implementation, you would inspect the page structure
            # and extract the relevant data
            store_elements = soup.select(".store-location")
            
            for store in store_elements:
                location = {
                    "store_name": store.select_one(".store-name").text.strip(),
                    "address": store.select_one(".address").text.strip(),
                    "city": store.select_one(".city").text.strip(),
                    "state": store.select_one(".state").text.strip(),
                    "zip_code": store.select_one(".zip").text.strip(),
                    # Coordinates might be available in data attributes
                    "latitude": store.get("data-lat"),
                    "longitude": store.get("data-lng"),
                }
                locations.append(location)
        
        return locations
    
    def _scrape_with_playwright(self) -> List[Dict[str, Any]]:
        """Scrape store locations using Playwright for dynamic content"""
        locations = []
        
        # Example implementation using Playwright
        # This is a placeholder that demonstrates the approach
        with sync_playwright() as p:
            browser = p.chromium.launch(headless=True)
            page = browser.new_page()
            
            # Navigate to the store locator page
            page.goto(self.store_locator_url)
            
            # Example: Enter a zip code to search for stores
            page.fill("#store-search-input", "90210")
            page.click("#store-search-button")
            
            # Wait for results to load
            page.wait_for_selector(".store-list-item")
            
            # Extract store data from the page
            # In a real implementation, you would inspect the page structure
            # and extract the relevant data
            store_elements = page.query_selector_all(".store-list-item")
            
            for store in store_elements:
                # Example extraction logic
                store_name = store.query_selector(".store-name").inner_text()
                address = store.query_selector(".address").inner_text()
                city = store.query_selector(".city").inner_text()
                state = store.query_selector(".state").inner_text()
                zip_code = store.query_selector(".zip").inner_text()
                
                # Coordinates might be available in data attributes or from a map
                lat = store.get_attribute("data-lat")
                lng = store.get_attribute("data-lng")
                
                location = {
                    "store_name": store_name,
                    "address": address,
                    "city": city,
                    "state": state,
                    "zip_code": zip_code,
                    "latitude": lat,
                    "longitude": lng,
                }
                locations.append(location)
            
            browser.close()
        
        # For demonstration purposes, return some sample data
        # In a real implementation, this would be the actual scraped data
        sample_locations = [
            {
                "store_name": "AutoZone #1234",
                "address": "123 Main St",
                "city": "Beverly Hills",
                "state": "CA",
                "zip_code": "90210",
                "latitude": "34.0736",
                "longitude": "-118.4004"
            },
            {
                "store_name": "AutoZone #5678",
                "address": "456 Oak Ave",
                "city": "Beverly Hills",
                "state": "CA",
                "zip_code": "90211",
                # No coordinates - will use geocoding
            }
        ]
        
        return sample_locations