requirements.txt
playwright==1.40.0
selenium==4.15.2
pandas==2.1.3
requests==2.31.0
beautifulsoup4==4.12.2
geopy==2.4.1
python-dotenv==1.0.0